scripts/requirements.txt
coolname==1.1.0
names==0.3.0
psycopg2-binary==2.9.3
